started to move time operations to lib
